+++ man-roxygen/plotsil.R
@@ -1,0 +1,7 @@
+#' @describeIn <%=class%> Plots the average silhouette width when the
+#'   clusters are cut by a sequence of k numbers. The k value with the highest
+#'   value (marked in the plot) may be considered as the optimal number of
+#'   clusters.
+#' @param kSeq An integer vector containing the sequence that should be used for
+#'   average silhouette width calculation.
+#' @param pch,type Passed to \code{\link[graphics]{plot}}.