+++ tests/testthat/test-compounds_clust.R
@@ -54,6 +54,10 @@ test_that("plotting works", {
     # these don't play well with vdiffr...
     expect_plot(plot(compsClust, groupName = firstGroup))
     expect_plot(plotStructure(compsClust, groupName = firstGroup, cluster = 1))
+    
+    # expect_doppel("compounds-clust-sil", function() plotSilhouettes(compsClust, kSeq = 2:6, groupName = firstGroup))
+    # UNDONE: currently all clusters have only two compounds so can't plot silhouette which needs >2
+    expect_error(plotSilhouettes(compsClust, kSeq = 2:6, groupName = firstGroup))
 })
 
 test_that("reporting works", {